Removed Tuple package and refactored cartesian series when
