+++ lib/src/charts/data/core/cartesian/cartesian_data.dart
@@ -1,7 +1,7 @@
 import 'package:chart_it/chart_it.dart';
 import 'package:chart_it/src/animations/tweens.dart';
 import 'package:chart_it/src/charts/painters/cartesian/cartesian_painter.dart';
-import 'package:chart_it/src/extensions/validators.dart';
+import 'package:chart_it/src/extensions/primitives.dart';
 import 'package:flutter/material.dart';
 
 /// Callback to construct a CartesianPainter for provided CartesianSeries Type
@@ -24,14 +24,28 @@ enum CartesianChartAlignment {
 enum CartesianChartOrientation { vertical, horizontal }
 
 abstract class CartesianSeries {
-  // T when<T>({required T Function() barSeries}) {
-  //   switch (runtimeType) {
-  //     case BarSeries:
-  //       return barSeries();
-  //     default:
-  //       throw TypeError();
-  //   }
-  // }
+  T when<T>({
+    required T Function() onBarSeries,
+  }) {
+    switch (runtimeType) {
+      case BarSeries:
+        return onBarSeries();
+      default:
+        throw TypeError();
+    }
+  }
+
+  T maybeWhen<T>({
+    T Function()? onBarSeries,
+    required T Function() orElse,
+  }) {
+    switch (runtimeType) {
+      case BarSeries:
+        return onBarSeries?.call() ?? orElse();
+      default:
+        throw TypeError();
+    }
+  }
 }
 
 List<Tween<CartesianSeries>>? toCartesianTweens(
@@ -43,8 +57,7 @@ List<Tween<CartesianSeries>>? toCartesianTweens(
         current == null || current.runtimeType != target.runtimeType
             ? null
             : current;
-    return whereSeries(
-      target.runtimeType,
+    return target.when(
       onBarSeries: () {
         return BarSeriesTween(
           begin: currentValue.asOrNull<BarSeries>(),
