+++ lib/src/extensions/data_conversions.dart
@@ -6,7 +6,6 @@ import 'package:chart_it/src/charts/data/bars/bar_series.dart';
 import 'package:chart_it/src/charts/data/bars/multi_bar.dart';
 import 'package:chart_it/src/charts/data/bars/simple_bar.dart';
 import 'package:chart_it/src/charts/data/core/cartesian/cartesian_data.dart';
-import 'package:chart_it/src/extensions/validators.dart';
 
 extension YValueGetter on BarGroup {
   // Helper method to strip group data into data object with raw values
@@ -26,8 +25,7 @@ extension Iterators on List<CartesianSeries> {
   int maxIterations() {
     var count = 0;
     forEach((series) {
-      whereSeries(
-        series.runtimeType,
+      series.when(
         onBarSeries: () {
           count = max(count, (series as BarSeries).barData.length);
         },
