+++ lib/src/extensions/primitives.dart
@@ -1,3 +1,12 @@
+extension AsExtension on Object? {
+  X as<X>() => this as X;
+
+  X? asOrNull<X>() {
+    var currObject = this;
+    return currObject is X ? currObject : null;
+  }
+}
+
 extension ListTransforms<T> on List<T> {
   T get(int index, T defaultValue) {
     if (index >= length) {
