+++ pubspec.yaml
@@ -11,7 +11,6 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  tuple: ^2.0.1
   equatable: ^2.0.5
   vector_math: ^2.1.4
